yolov8_urban_trees/requirements.txt
ultralytics>=8.2.0
rasterio>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.30.0
pillow>=9.0.0
